feat(timestamps): update timestamps when doing `$set` on a child path
Re: #4412
