+++ lib/helpers/update/applyTimestampsToUpdate.js
@@ -12,9 +12,9 @@ module.exports = applyTimestampsToUpdate;
  * ignore
  */
 
-function applyTimestampsToUpdate(createdAt, updatedAt, currentUpdate, options, inPlace) {
+function applyTimestampsToUpdate(createdAt, updatedAt, currentUpdate, options) {
   const now = new Date();
-  const updates = inPlace ? currentUpdate : {};
+  const updates = currentUpdate;
   let _updates = updates;
   const overwrite = get(options, 'overwrite', false);
   const timestamps = get(options, 'timestamps', true);
