+++ lib/model.js
@@ -2908,9 +2908,7 @@ Model.bulkWrite = function(ops, options, callback) {
           if (this.schema.$timestamps != null) {
             const createdAt = this.schema.$timestamps.createdAt;
             const updatedAt = this.schema.$timestamps.updatedAt;
-            const inPlace = true;
-            applyTimestampsToUpdate(createdAt, updatedAt, op['update'], {},
-              inPlace);
+            applyTimestampsToUpdate(createdAt, updatedAt, op['update'], {});
           }
         } catch (error) {
           return callback(error, null);
@@ -2936,9 +2934,7 @@ Model.bulkWrite = function(ops, options, callback) {
           if (this.schema.$timestamps != null) {
             const createdAt = this.schema.$timestamps.createdAt;
             const updatedAt = this.schema.$timestamps.updatedAt;
-            const inPlace = true;
-            applyTimestampsToUpdate(createdAt, updatedAt, op['update'], {},
-              inPlace);
+            applyTimestampsToUpdate(createdAt, updatedAt, op['update'], {});
           }
         } catch (error) {
           return callback(error, null);
