+++ lib/schema.js
@@ -8,9 +8,11 @@ const EventEmitter = require('events').EventEmitter;
 const Kareem = require('kareem');
 const SchemaType = require('./schematype');
 const VirtualType = require('./virtualtype');
+const applyTimestampsToChildren = require('./helpers/update/applyTimestampsToChildren');
 const applyTimestampsToUpdate = require('./helpers/update/applyTimestampsToUpdate');
 const get = require('lodash.get');
 const getIndexes = require('./helpers/schema/getIndexes');
+const handleTimestampOption = require('./helpers/schema/handleTimestampOption');
 const merge = require('./helpers/schema/merge');
 const mpath = require('mpath');
 const readPref = require('./driver').get().ReadPreference;
@@ -23,6 +25,8 @@ let MongooseTypes;
 const allMiddleware = require('./helpers/query/applyQueryMiddleware').
   middlewareFunctions.
   concat(require('./helpers/model/applyHooks').middlewareFunctions);
+
+let id = 0;
 
 /**
  * Schema constructor.
@@ -91,6 +95,8 @@ function Schema(obj, options) {
   this.query = {};
   this.childSchemas = [];
   this.plugins = [];
+  // For internal debugging. Do not use this to try to save a schema in MDB.
+  this.$id = ++id;
 
   this.s = {
     hooks: new Kareem()
@@ -116,9 +122,7 @@ function Schema(obj, options) {
     this.add(_obj);
   }
 
-  if (this.options.timestamps) {
-    this.setupTimestamp(this.options.timestamps);
-  }
+  this.setupTimestamp(this.options.timestamps);
 }
 
 /*!
@@ -529,6 +533,10 @@ Schema.prototype.path = function(path, obj) {
 
   this.paths[path] = Schema.interpretAsType(path, obj, this.options);
   const schemaType = this.paths[path];
+  if (schemaType.schema != null) {
+    schemaType.schema.$schemaType = schemaType;
+  }
+  schemaType.$parentSchema = this;
 
   if (schemaType.$isSchemaMap) {
     // Maps can have arbitrary keys, so `$*` is internal shorthand for "any key"
@@ -822,7 +830,10 @@ Schema.prototype.hasMixedParent = function(path) {
  * @api private
  */
 Schema.prototype.setupTimestamp = function(timestamps) {
-  if (timestamps == null) {
+  const childHasTimestamp = this.childSchemas.
+    find(s => s.schema.options.timestamps != null);
+
+  if (timestamps == null && !childHasTimestamp) {
     return;
   }
 
@@ -882,116 +893,11 @@ Schema.prototype.setupTimestamp = function(timestamps) {
 
   function _setTimestampsOnUpdate(next) {
     applyTimestampsToUpdate(createdAt, updatedAt, this.getUpdate(),
-      this.options, true);
+      this.options, this.schema);
     applyTimestampsToChildren(this);
     next();
   }
 };
-
-/*!
- * ignore
- */
-
-function handleTimestampOption(arg, prop) {
-  if (typeof arg === 'boolean') {
-    return prop;
-  }
-  if (typeof arg[prop] === 'boolean') {
-    return arg[prop] ? prop : null;
-  }
-  if (!(prop in arg)) {
-    return prop;
-  }
-  return arg[prop];
-}
-
-/*!
- * ignore
- */
-
-function applyTimestampsToChildren(query) {
-  const now = new Date();
-  const update = query.getUpdate();
-  const keys = Object.keys(update);
-  let key;
-  const schema = query.model.schema;
-  let len;
-  let createdAt;
-  let updatedAt;
-  let timestamps;
-  let path;
-
-  const hasDollarKey = keys.length && keys[0].charAt(0) === '$';
-
-  if (hasDollarKey) {
-    if (update.$push) {
-      for (key in update.$push) {
-        const $path = schema.path(key);
-        if (update.$push[key] &&
-            $path &&
-            $path.$isMongooseDocumentArray &&
-            $path.schema.options.timestamps) {
-          timestamps = $path.schema.options.timestamps;
-          createdAt = handleTimestampOption(timestamps, 'createdAt');
-          updatedAt = handleTimestampOption(timestamps, 'updatedAt');
-          if (update.$push[key].$each) {
-            update.$push[key].$each.forEach(function(subdoc) {
-              if (updatedAt != null) {
-                subdoc[updatedAt] = now;
-              }
-              if (createdAt != null) {
-                subdoc[createdAt] = now;
-              }
-            });
-          } else {
-            if (updatedAt != null) {
-              update.$push[key][updatedAt] = now;
-            }
-            if (createdAt != null) {
-              update.$push[key][createdAt] = now;
-            }
-          }
-        }
-      }
-    }
-    if (update.$set) {
-      for (key in update.$set) {
-        path = schema.path(key);
-        if (!path) {
-          continue;
-        }
-        if (Array.isArray(update.$set[key]) && path.$isMongooseDocumentArray) {
-          len = update.$set[key].length;
-          timestamps = schema.path(key).schema.options.timestamps;
-          if (timestamps) {
-            createdAt = handleTimestampOption(timestamps, 'createdAt');
-            updatedAt = handleTimestampOption(timestamps, 'updatedAt');
-            for (let i = 0; i < len; ++i) {
-              if (updatedAt != null) {
-                update.$set[key][i][updatedAt] = now;
-              }
-              if (createdAt != null) {
-                update.$set[key][i][createdAt] = now;
-              }
-            }
-          }
-        } else if (update.$set[key] && path.$isSingleNested) {
-          timestamps = schema.path(key).schema.options.timestamps;
-          if (timestamps) {
-            createdAt = handleTimestampOption(timestamps, 'createdAt');
-            updatedAt = handleTimestampOption(timestamps, 'updatedAt');
-            if (updatedAt != null) {
-              update.$set[key][updatedAt] = now;
-            }
-            if (createdAt != null) {
-              update.$set[key][createdAt] = now;
-            }
-          }
-        }
-      }
-    }
-  }
-}
 
 /*!
  * ignore
